Modified CVARINFO, ACS Script, and Decorate/Weapons/Misc.txt. Tweaked the characteristics on how Gun Mode processes the current inventory in case the user decides to change it in-game.
The following process happens for each new game:

1. Upon starting, if the player has Gun Mode enabled at first or decides
to enable it in-game later, it will give the appropriate ballistic weapons
to the player based on the superpowers the player has, plus the ammunition
for them (The math used for this is that we take the max energy and divide
it by the max ammo for each weapon to get a number to divide the current
energy by to get the appropriate ammo for each gun. i.e. For the shotgun,
600/50=12, so if we have 300 energy, then we get 300/12=25 shells. The
number is then truncated if decimal places exist.).

2. If the player then decides to disable gun mode in-game, the guns
themselves are then taken away from the player.

3. Finally, if the player decides to enable gun mode once again, then it's
the same process as #1, except that it does not give the player ammo for
each weapon.

4. #2 and #3 repeat as the player goes back and forth between modes until
the game ends.